Move tokenPrices service to backend service, create type file for shared types
Move getLPShareCode to balances helper
